apply ode_analyer to analyze Lorenz system, which has fixed points, limit cycles and chaotic solutions.
